pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "searchgoat-hex"
version = "0.5.0"
description = "Query Cribl Search from Hex notebooks. Get DataFrames."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "JR", email = "jr@ross.moda" }
]
keywords = ["cribl", "search", "hex", "dataframe", "pandas", "observability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "requests>=2.28",
    "pandas>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/hackish-pub/searchgoat-hex"
Documentation = "https://github.com/hackish-pub/searchgoat-hex#readme"
Repository = "https://github.com/hackish-pub/searchgoat-hex"

[tool.hatch.build.targets.wheel]
packages = ["src/searchgoat_hex"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
